requirements.txt
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
et-xmlfile==1.0.1
idna==2.10
isodate==0.6.0
jdcal==1.4.1
modin==0.8.0
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.0.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
wincertstore==0.2
